Move utils and const in src
